stats-bot/pyproject.toml
[project]
name = "stats-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "discord-py>=2.6.4",
    "duckdb>=1.4.1",
    "pydantic-ai-slim[openai]>=1.9.1",
    "pydantic-settings>=2.11.0",
    "pytz>=2025.2",
]

[dependency-groups]
dev = [
    "ruff>=0.14.3",
]
